[ui,data] Create ConfigItem, ConfigList
Adds a ConfigItem class for defining a config item with it's
relationships to other config items, as well as it's currently selected
state.

Additionally, adds a VConfigItem class which will be used for the visual
representation of a ConfigItem, and a ConfigList which will be used to
list a group of ConfigItems.

NOTE: main.py does not work with these changes, and will be fixed in
a future commit.